reorganized visualization subpackage
