+++ core/eolearn/core/eodata.py
@@ -626,7 +626,7 @@ class EOPatch:
         """
         # pylint: disable=import-outside-toplevel,raise-missing-from
         try:
-            from eolearn.visualization import EOPatchVisualization
+            from eolearn.visualization.eopatch import EOPatchVisualization
         except ImportError:
             raise RuntimeError('Subpackage eo-learn-visualization has to be installed with an option [FULL] in order '
                                'to use plot method')
