+++ core/eolearn/core/eoexecution.py
@@ -323,7 +323,7 @@ class EOExecutor:
         """
         # pylint: disable=import-outside-toplevel,raise-missing-from
         try:
-            from eolearn.visualization import EOExecutorVisualization
+            from eolearn.visualization.eoexecutor import EOExecutorVisualization
         except ImportError:
             raise RuntimeError('Subpackage eo-learn-visualization has to be installed in order to create EOExecutor '
                                'reports')
