+++ visualization/eolearn/visualization/__init__.py
@@ -2,12 +2,4 @@
 A collection of visualization utilities
 """
 
-from .eoworkflow_visualization import EOWorkflowVisualization
-from .eoexecutor_visualization import EOExecutorVisualization
-
-try:
-    from .eopatch_visualization import EOPatchVisualization
-except ImportError:
-    pass
-
 __version__ = '1.0.0'
